feat(camera): add pause event if click 'esc' during preview on beam series, bb2
